separate unit/integration Asset tests
